Rename cli._corecommand module to cli.base
