+++ src/repobee_plug/cli/__init__.py
@@ -13,8 +13,8 @@ from repobee_plug.cli.args import (
 )
 from repobee_plug.cli.commandmarkers import Command, CommandExtension
 
-from repobee_plug.cli._corecommand import _CoreCommand
-from repobee_plug.cli._corecommand import BaseParser
+from repobee_plug.cli.base import _CoreCommand
+from repobee_plug.cli.base import BaseParser
 
 CoreCommand = _CoreCommand()
